# Conflicts:
#	Compiler/Applications/LSP/LSPServer.Core/Server.cs

+++ Compiler/Commons/ValueObjects/LongValueObject.cs
@@ -1,4 +1,4 @@
 namespace KSPCompiler.Commons.ValueObjects;
 
 // ReSharper disable once UnusedType.Global
-public abstract record LongValueObject( int Value ) : ValueObject<int>( Value );
+public abstract record LongValueObject( long Value ) : ValueObject<long>( Value );
