+++ Compiler/Commons/ValueObjects/ULongValueObject.cs
@@ -1,4 +1,4 @@
 namespace KSPCompiler.Commons.ValueObjects;
 
 // ReSharper disable once UnusedType.Global
-public abstract record ULongValueObject( uint Value ) : ValueObject<uint>( Value );
+public abstract record ULongValueObject( ulong Value ) : ValueObject<ulong>( Value );
